+++ django_lightningkite/notifications/channels/channel.py
@@ -1,6 +1,6 @@
 class Channel():
 
-    def send(notifiable, notification):
+    def send(notifiable, notification, *args, **kwargs):
         """
         Send the Given Notifiction
         """