+++ django_lightningkite/notifications/channels/console_channel.py
@@ -7,7 +7,7 @@ from .. import SUCCESS, FAILED
 class ConsoleChannel(Channel):
 
     @staticmethod
-    def send(notifiable, notification):
+    def send(notifiable, notification, *args, **kwargs):
         """
         Send the Given Notifiction
         """
@@ -17,6 +17,6 @@ class ConsoleChannel(Channel):
         """
         sending.send(sender=ConsoleChannel, notifiable=notifiable, notification=notification)
         # TODO make thread safe: https://github.com/django/django/blob/master/django/core/mail/backends/console.py
-        message = notification.to_console(notifiable)
+        message = notification.to_console(notifiable, *args, **kwargs)
         sys.stdout.write('{}\n'.format(message))
         sent.send(sender=ConsoleChannel, notifiable=notifiable, notification=notification, message=message, status=SUCCESS)