+++ django_lightningkite/notifications/utils.py
@@ -2,6 +2,7 @@ from django.template import loader
 from django.core.mail import EmailMultiAlternatives
 from pynliner import Pynliner
 
+
 def email_from_template(html_template, txt_template, context, from_email):
     html_email = loader.render_to_string(html_template, context)
     html_email = Pynliner().from_string(html_email).run()
@@ -15,5 +16,3 @@ def email_from_template(html_template, txt_template, context, from_email):
     msg = EmailMultiAlternatives(**details)
     msg.attach_alternative(html_email, 'text/html')
     return msg
-
-